+++ packages/tools/src/enums/Events.ts
@@ -39,6 +39,13 @@ enum Events {
 
   VOLUMECROPPING_TOOL_CHANGED = 'CORNERSTONE_TOOLS_VOLUMECROPPING_TOOL_CHANGED',
 
+  /**
+   * Triggers on the eventTarget when a stack prefetch has been completed (i.e., the indicesToRequest list
+   * has been emptied). It is triggered both by the stackPrefetch and stackContextPrefetch mechanism.
+   * See what event detail is included in {@link EventTypes.StackPrefetchCompleteEventDetail | Stack Prefetch Complete Event Detail}.
+   */
+  STACK_PREFETCH_COMPLETE = 'CORNERSTONE_TOOLS_STACK_PREFETCH_COMPLETE',
+
   ///////////////////////////////////////
   //            Annotations
   ///////////////////////////////////////