+++ packages/tools/src/tools/CrosshairsTool.ts
@@ -4,7 +4,7 @@ import vtkMatrixBuilder from '@kitware/vtk.js/Common/Core/MatrixBuilder';
 
 import { AnnotationTool } from './base';
 
-import type { Types } from '@cornerstonejs/core';
+import { getRenderingEngine, type Types } from '@cornerstonejs/core';
 import {
   getEnabledElementByIds,
   getEnabledElement,
@@ -158,6 +158,11 @@ class CrosshairsTool extends AnnotationTool {
         filterActorUIDsToSetSlabThickness: [],
         // blend mode for slabThickness modifications
         slabThicknessBlendMode: Enums.BlendModes.MAXIMUM_INTENSITY_BLEND,
+        centerPoint: {
+          enabled: false,
+          color: 'rgba(255, 255, 0, 0.5)',
+          size: 2,
+        },
         mobile: {
           enabled: false,
           opacity: 0.8,
@@ -412,13 +417,57 @@ class CrosshairsTool extends AnnotationTool {
 
   setToolCenter(toolCenter: Types.Point3, suppressEvents = false): void {
     // prettier-ignore
-    this.toolCenter = toolCenter;
     const viewportsInfo = this._getViewportsInfo();
 
-    // assuming all viewports are in the same rendering engine
-    triggerAnnotationRenderForViewportIds(
-      viewportsInfo.map(({ viewportId }) => viewportId)
-    );
+    viewportsInfo.map(({ renderingEngineId, viewportId }) => {
+      const renderingEngine = getRenderingEngine(renderingEngineId);
+
+      const viewport = renderingEngine.getViewport(viewportId);
+      const camera = viewport.getCamera();
+      const { focalPoint, position, viewPlaneNormal } = camera;
+
+      // Calculate the delta between the current camera focal point and the new tool center
+      const delta = [
+        toolCenter[0] - focalPoint[0],
+        toolCenter[1] - focalPoint[1],
+        toolCenter[2] - focalPoint[2],
+      ];
+
+      // Project this vector onto the view plane normal.
+      // This isolates the component of the movement that corresponds to the "scroll" (slice change).
+      const scroll =
+        delta[0] * viewPlaneNormal[0] +
+        delta[1] * viewPlaneNormal[1] +
+        delta[2] * viewPlaneNormal[2];
+
+      const scrollDelta = [
+        scroll * viewPlaneNormal[0],
+        scroll * viewPlaneNormal[1],
+        scroll * viewPlaneNormal[2],
+      ];
+
+      // Apply this "scroll" to the position and focal point of the camera.
+      const newFocalPoint: Types.Point3 = [
+        focalPoint[0] + scrollDelta[0],
+        focalPoint[1] + scrollDelta[1],
+        focalPoint[2] + scrollDelta[2],
+      ];
+      const newPosition: Types.Point3 = [
+        position[0] + scrollDelta[0],
+        position[1] + scrollDelta[1],
+        position[2] + scrollDelta[2],
+      ];
+
+      viewport.setCamera({
+        focalPoint: newFocalPoint,
+        position: newPosition,
+      });
+
+      viewport.render();
+    });
+
+    this.toolCenter = toolCenter;
+
     if (!suppressEvents) {
       triggerEvent(eventTarget, Events.CROSSHAIR_TOOL_CENTER_CHANGED, {
         toolGroupId: this.toolGroupId,
@@ -1501,6 +1550,31 @@ class CrosshairsTool extends AnnotationTool {
         referenceColorCoordinates as Types.Point2,
         circleRadius,
         { color, fill: color }
+      );
+    }
+
+    if (this.configuration.centerPoint?.enabled) {
+      const defaultColor = 'rgba(255, 255, 0, 0.5)';
+      const defaultSize = 2;
+      const maxAllowedSize = 5;
+
+      const centerPointColor =
+        this.configuration.centerPoint.color || defaultColor;
+      const centerPointSize = Math.min(
+        this.configuration.centerPoint.size || defaultSize,
+        maxAllowedSize
+      );
+
+      drawCircleSvg(
+        svgDrawingHelper,
+        annotationUID,
+        'centerPoint',
+        crosshairCenterCanvas as Types.Point2,
+        centerPointSize,
+        {
+          color: centerPointColor,
+          fill: centerPointColor,
+        }
       );
     }